pyproject.toml
[tool.poetry]
name = "fractal-matrix-client"
version = "0.1.0"
description = ""
authors = ["Mo Balaa <balaa@fractalnetworks.co>"]
readme = "README.md"
packages = [
 {include = "fractal"} ]

[tool.poetry.dependencies]
python = "^3.10"
appdirs = "^1.4.4"
matrix-nio = "^0.22.1"
sh = "^2.0.6"
aiofiles = "^23.1.0"
PyYAML = "^6.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.plugins."fractal.plugins"]
"login" = "fractal.matrix.controllers.login"